Move checkout instruction to utils
