+++ src/instructions/express_checkout.ts
@@ -1,14 +1,5 @@
-import { Buffer } from 'buffer';
 import type { Connection, TransactionSignature } from '@solana/web3.js';
-import {
-  PublicKey,
-  Transaction,
-  TransactionInstruction,
-  SYSVAR_CLOCK_PUBKEY,
-  SYSVAR_RENT_PUBKEY,
-  SystemProgram,
-} from '@solana/web3.js';
-import { TOKEN_PROGRAM_ID } from '../helpers/solana';
+import { PublicKey, Transaction } from '@solana/web3.js';
 import type { Result } from '../helpers/result';
 import { failure } from '../helpers/result';
 import type { WalletAdapter } from '../helpers/types';
@@ -16,8 +7,8 @@ import {
   awaitTransactionSignatureConfirmation,
   signAndSendTransaction,
 } from '../helpers/transaction';
-import { Instruction, InstructionData, InstructionType } from '../helpers/instruction';
-import { MAX_SEED_LEN, PDA_SEED } from '../helpers/constants';
+import { InstructionType } from '../helpers/instruction';
+import { makeExpressCheckoutTransaction } from './utils';
 
 interface ExpressCheckoutParams {
   amount: number;
@@ -34,25 +25,6 @@ interface ExpressCheckoutParams {
   wallet: WalletAdapter;
 }
 
-export const getOrderAccountPubkey = async (
-  orderId: string,
-  wallet_pk: PublicKey,
-  programIdKey: PublicKey
-): Promise<PublicKey> => {
-  return PublicKey.createWithSeed(wallet_pk, orderId.slice(0, MAX_SEED_LEN), programIdKey);
-};
-
-export const getSellerTokenAccountPubkey = async (
-  orderAccount: PublicKey,
-  mint: PublicKey,
-  programIdKey: PublicKey
-): Promise<[PublicKey, number]> => {
-  return PublicKey.findProgramAddress(
-    [orderAccount.toBytes(), TOKEN_PROGRAM_ID.toBytes(), mint.toBytes()],
-    programIdKey
-  );
-};
-
 export const expressCheckout = async (
   params: ExpressCheckoutParams
 ): Promise<Result<TransactionSignature>> => {
@@ -60,6 +32,7 @@ export const expressCheckout = async (
     amount,
     buyerTokenAccount,
     connection,
+    data,
     merchantAccount,
     mint,
     programOwnerAccount,
@@ -72,48 +45,23 @@ export const expressCheckout = async (
   if (!wallet.publicKey) {
     return failure(new Error('Wallet not connected'));
   }
-  const data = params.data || null;
   const programIdKey = new PublicKey(thisProgramId);
-  const orderAccount = await getOrderAccountPubkey(orderId, wallet.publicKey, programIdKey);
-  const promiseResults = await Promise.all([
-    PublicKey.findProgramAddress([Buffer.from(PDA_SEED)], programIdKey),
-    getSellerTokenAccountPubkey(orderAccount, mint, programIdKey),
-  ]);
-  const pda = promiseResults[0];
-  const sellerTokenAccount = promiseResults[1];
-
   const transaction = new Transaction({ feePayer: wallet.publicKey });
 
   try {
     transaction.add(
-      new TransactionInstruction({
+      await makeExpressCheckoutTransaction({
+        amount,
+        buyerTokenAccount,
+        data,
+        merchantAccount,
+        mint,
+        orderId,
         programId: programIdKey,
-        keys: [
-          { pubkey: wallet.publicKey, isSigner: true, isWritable: true },
-          { pubkey: orderAccount, isSigner: false, isWritable: true },
-          { pubkey: merchantAccount, isSigner: false, isWritable: false },
-          { pubkey: sellerTokenAccount[0], isSigner: false, isWritable: true },
-          { pubkey: buyerTokenAccount, isSigner: false, isWritable: true },
-          { pubkey: programOwnerAccount, isSigner: false, isWritable: true },
-          { pubkey: sponsorAccount, isSigner: false, isWritable: true },
-          { pubkey: mint, isSigner: false, isWritable: false },
-          { pubkey: pda[0], isSigner: false, isWritable: false },
-          { pubkey: TOKEN_PROGRAM_ID, isSigner: false, isWritable: false },
-          { pubkey: SystemProgram.programId, isSigner: false, isWritable: false },
-          { pubkey: SYSVAR_CLOCK_PUBKEY, isSigner: false, isWritable: false },
-          { pubkey: SYSVAR_RENT_PUBKEY, isSigner: false, isWritable: false },
-        ],
-        data: new Instruction({
-          instruction: InstructionType.ExpressCheckout,
-          [InstructionType.ExpressCheckout]: new Uint8Array(
-            new InstructionData(InstructionType.ExpressCheckout, {
-              amount,
-              order_id: orderId,
-              secret,
-              data,
-            }).encode()
-          ),
-        }).encode(),
+        programOwnerAccount,
+        secret,
+        sponsorAccount,
+        signer: wallet.publicKey,
       })
     );
   } catch (error) {
